Use .items() instead of .iteritems() for Python 3 compatibility
Cython actually handles iteritems() even on Python 3 by (presumably)
converting it into .items() when called on a dictionary.

However, OrderedDict (and presumably other custom dict subclasses) don't
have an iteritems method, so they cannot be inserted as an argument
to setncatts.

This change switches two uses of iteritems() to items(). I didn't
bother to avoid constructing the intermediate list in Python 2, because
it seems unlikely to me that the slight performance difference would affect
anyone, unless they have an outrageous number of attributes that they are
setting at once. But if you like, I could write a small cross-version
compatibility function, so that a temporary list is not created.